Implementado Knockback, mejor interfaz, otras cosas mínimas.

+++ Proyecto2hu/core/src/com/mygdx/game/MoveByPixel.java
@@ -6,12 +6,12 @@ import com.badlogic.gdx.math.Rectangle;
 public class MoveByPixel implements IMovement{
 
 	@Override
-	public void moveLeft(Rectangle hitbox, int vel) {
+	public void moveLeft(Rectangle hitbox, float vel) {
 		hitbox.x -= vel * Gdx.graphics.getDeltaTime();
 	}
 
 	@Override
-	public void moveRight(Rectangle hitbox, int vel) {
+	public void moveRight(Rectangle hitbox, float vel) {
 		hitbox.x += vel * Gdx.graphics.getDeltaTime();
 	}
 }
