+++ Proyecto2hu/core/src/com/mygdx/game/MoveVectorial.java
@@ -6,12 +6,12 @@ import com.badlogic.gdx.math.Rectangle;
 public class MoveVectorial implements IMovement {
 
 	@Override
-	public void moveLeft(Rectangle hitbox, int vel) {
+	public void moveLeft(Rectangle hitbox, float vel) {
 		hitbox.x -= vel * 2 * Gdx.graphics.getDeltaTime();
 	}
 
 	@Override
-	public void moveRight(Rectangle hitbox, int vel) {
+	public void moveRight(Rectangle hitbox, float vel) {
 		hitbox.x += vel * 2 * Gdx.graphics.getDeltaTime();
 	}
 }
